Fix and organize burn tests
